clean up schema utils, add GetURL
