+++ enterprise/cmd/frontend/internal/auth/githuboauth/provider.go
@@ -23,10 +23,7 @@ import (
 const sessionKey = "githuboauth@0"
 
 func parseProvider(p *schema.GitHubAuthProvider, db dbutil.DB, sourceCfg schema.AuthProviders) (provider *oauth.Provider, messages []string) {
-	rawURL := p.Url
-	if rawURL == "" {
-		rawURL = schema.DefaultGitHubURL
-	}
+	rawURL := p.GetURL()
 	parsedURL, err := url.Parse(rawURL)
 	if err != nil {
 		messages = append(messages, fmt.Sprintf("Could not parse GitHub URL %q. You will not be able to login via this GitHub instance.", rawURL))
