--- schema/defaults.go
@@ -1,4 +1,0 @@
-package schema
-
-// DefaultGitHubURL is the default GitHub instance that configuration points to.
-const DefaultGitHubURL = "https://github.com/"
