+++ schema/gitlab_util.go
@@ -1,17 +1,5 @@
 package schema
 
-func (c *GitHubConnection) SetRepos(all bool, repos []string) error {
-	if all {
-		c.RepositoryQuery = []string{"affiliated"}
-		c.Repos = nil
-		return nil
-	} else {
-		c.RepositoryQuery = []string{}
-	}
-	c.Repos = repos
-	return nil
-}
-
 func (c *GitLabConnection) SetRepos(all bool, repos []string) error {
 	if all {
 		c.ProjectQuery = []string{"projects?membership=true&archived=no"}
